Pascal,
I took a quick look.  One thing I noticed is that it seems like the Abstract needs to be updated.   For example it mentions “extends RFC 4291”.     The last paragraph of the Introduction might be a better abstract for this document:

This document presents an architecture for IPv6 access networks that decouples the lower-layer concepts of links and broadcast domains from the IP concepts of Links and Subnets, and limits the use and impact of broadcast. This architecture is suitable for IPv6 over any network, including modern Non-Broadcast MultiAccess (NBMA) and non-transitive Point-to-Multipoint (P2MP) networks. A study of the issues with ND-Classic over wireless media is presented, and a framework to solve those issues within the new architecture is proposed.

Bob